requirements.txt
asgiref==3.3.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.8
Pillow==8.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycparser==2.20
pyparsing==2.4.7
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
